@@ -1,14 +1,9 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.8'
-gem 'rake', '10.1.1'
+gem 'rails', '5.1.3'
 
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+gem 'sqlite3'
 
-#gem 'sqlite3'
-
-gem 'mysql2'
 gem 'haml-rails'
 gem 'devise'
 
@@ -19,24 +14,23 @@ gem 'nokogiri'
 gem "twitter-bootstrap-rails"
 #gem 'twitter-bootstrap-rails', :git => 'git://github.com/seyhunak/twitter-bootstrap-rails.git'
 
-#gem "squeel"
-gem 'squeel'
+# gem 'squeel'
 gem 'thin'
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
   gem 'less-rails'
-  gem 'sass-rails', '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails'
+  gem 'coffee-rails'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   gem 'therubyracer', :platforms => :ruby
 
-  gem 'uglifier', '>= 1.0.3'
+  gem 'uglifier'
 
-  gem 'sass-rails', '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails'
+  gem 'coffee-rails'
 end
 
 group :development, :test do
@@ -46,7 +40,7 @@ group :development, :test do
   #gem 'binding_of_caller'
   gem 'meta_request'
 
-  gem 'rspec-rails', '~> 2.0'
+  gem 'rspec-rails'
 end
 
 group :production do
@@ -70,3 +64,5 @@ gem 'jquery-rails'
 
 # To use debugger
 # gem 'debugger'
+
+gem 'listen'
