@@ -1,4 +1,4 @@
-class AddFieldToUsers < ActiveRecord::Migration
+class AddFieldToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :role, :integer
   end
