@@ -1,4 +1,4 @@
-class AddRatingFieldToPointers < ActiveRecord::Migration
+class AddRatingFieldToPointers < ActiveRecord::Migration[4.2]
   def change
     add_column :pointers, :rating, :integer, :default => 0
   end
