@@ -1,4 +1,4 @@
-class AddIndexToPointers < ActiveRecord::Migration
+class AddIndexToPointers < ActiveRecord::Migration[4.2]
   def change
     add_index :pointers, [:latitude, :longitude], unique: true
   end
