@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -9,54 +8,52 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20141026130335) do
+ActiveRecord::Schema.define(version: 20141026130335) do
 
-  create_table "desires", :force => true do |t|
-    t.integer  "pointer_id"
-    t.integer  "user_id"
-    t.integer  "stat"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+  create_table "desires", force: :cascade do |t|
+    t.integer "pointer_id"
+    t.integer "user_id"
+    t.integer "stat"
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "pointers", :force => true do |t|
-    t.float    "latitude"
-    t.float    "longitude"
-    t.string   "description"
-    t.text     "full_desc"
-    t.boolean  "gmaps"
-    t.datetime "created_at",                 :null => false
-    t.datetime "updated_at",                 :null => false
-    t.date     "rec_date"
-    t.integer  "rating",      :default => 0
+  create_table "pointers", force: :cascade do |t|
+    t.float "latitude"
+    t.float "longitude"
+    t.string "description"
+    t.text "full_desc"
+    t.boolean "gmaps"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.date "rec_date"
+    t.integer "rating", default: 0
+    t.index ["latitude", "longitude"], name: "index_pointers_on_latitude_and_longitude", unique: true
   end
 
-  add_index "pointers", ["latitude", "longitude"], :name => "index_pointers_on_latitude_and_longitude", :unique => true
-
-  create_table "users", :force => true do |t|
-    t.string   "email",                  :default => "", :null => false
-    t.string   "encrypted_password",     :default => "", :null => false
-    t.string   "reset_password_token"
+  create_table "users", force: :cascade do |t|
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          :default => 0
+    t.integer "sign_in_count", default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
-    t.string   "confirmation_token"
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
+    t.string "confirmation_token"
     t.datetime "confirmed_at"
     t.datetime "confirmation_sent_at"
-    t.string   "unconfirmed_email"
-    t.datetime "created_at",                             :null => false
-    t.datetime "updated_at",                             :null => false
-    t.integer  "role"
+    t.string "unconfirmed_email"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.integer "role"
+    t.index ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
-
-  add_index "users", ["confirmation_token"], :name => "index_users_on_confirmation_token", :unique => true
-  add_index "users", ["email"], :name => "index_users_on_email", :unique => true
-  add_index "users", ["reset_password_token"], :name => "index_users_on_reset_password_token", :unique => true
 
 end
