@@ -1,6 +1,6 @@
 class Pointer < ActiveRecord::Base
 
-  attr_accessible :description, :full_desc, :gmaps, :latitude, :longitude, :rec_date
+  # attr_accessible :description, :full_desc, :gmaps, :latitude, :longitude, :rec_date
   #acts_as_gmappable :process_geocoding => false
 
   has_many :desires
